requirements.txt
# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# Telegram bot
aiogram>=3.15.0

# Env and logging
python-dotenv>=1.0.0
